Revert "unit test and dom tests"
