+++ .github/workflows/puppeteer.yml
@@ -15,8 +15,5 @@ jobs:
           node-version: '20'
 
       - run: npm install
-      
-      - name: Standard Page Movement Test
-        run: npm test ./tests/puppeteer.test.js
-      - name: Theme Toggle Test
-        run: npm test ./tests/theme-toggle.test.js
+
+      - run: npm test ./tests/puppeteer-demo.test.js
