+++ .github/workflows/tests.yml
@@ -21,11 +21,3 @@ jobs: # the things being executed
         run: npm test ./tests/core-blast-deck.test.js
       - name: Arm Workout Deck Test
         run: npm test ./tests/arm-workout-deck.test.js
-      - name: Create Deck Dom Unit Test
-        run: npm test ./tests/dom-functions-create-deck.test.js
-      - name: Create Deck / Filter / Deck Logic Unit Test
-        run: npm test ./tests/unit-test-create-deck.test.js
-      - name: Delete Deck Unit Test
-        run: npm test ./tests/delete-deck.test.js
-      - name: Shuffle Unit Test
-        run: npm test ./tests/shuffle.test.js
