+++ package.json
@@ -36,7 +36,6 @@
     "eslint-plugin-jsdoc": "^50.6.11",
     "globals": "^16.1.0",
     "jest": "^29.7.0",
-    "jest-environment-jsdom": "^30.0.0-beta.3",
     "jest-puppeteer": "^11.0.0",
     "jsdoc": "^4.0.4",
     "puppeteer": "^24.10.0"
