+++ tests/arm-workout-deck.test.js
@@ -56,6 +56,4 @@ describe('create arm workout deck test', () => {
     const workout3 = await getWorkoutName();
     expect(workout3).toBe("Shoulder Press");
   }, 90000);
-
-  
 });
