+++ tests/chest-day-deck.test.js
@@ -52,6 +52,4 @@ describe('create chest day deck test', () => {
     const workout3 = await getWorkoutName();
     expect(workout3).toBe("Incline Press");
   }, 60000);
-
- 
 });
